Move algorithm functions
